feat(task_sdk): Move assets.metadata to task_sdk.definitions.asset
